requirements.txt
streamlit
pandas
numpy
scikit-learn
spacy==3.7.4  # Specific version to ensure compatibility
nltk
transformers
python-docx
pdfminer.six
# For GCP integration (if you enable GCS fetch or Vertex AI calls)
# google-cloud-storage
# google-cloud-aiplatform